feat: add sendTransaction to signer interface & fix: move some functions
